requirements.txt
beautifulsoup4
requests
google-generativeai        
python-dotenv
google
google-genai
pandas
feedparser
